isPrivate logic in ExploreSpaces, isPrivate icon adjustements
