+++ src/main/topLevelPages/myDashboard/ExploreSpaces/ExploreSpacesTypes.ts
@@ -1,6 +1,7 @@
 import { Visual } from '../../../../domain/common/visual/Visual';
 import { SimpleContainerProps } from '../../../../core/container/SimpleContainer';
 import { Identifiable } from '../../../../core/utils/Identifiable';
+import { SpacePrivacyMode } from '../../../../core/apollo/generated/graphql-schema';
 
 export interface ExploreSpacesContainerEntities {
   spaces: SpaceWithParent[] | undefined;
@@ -28,6 +29,11 @@ interface ParentSpace extends Identifiable {
     displayName: string;
     avatar?: Visual;
     cardBanner?: Visual;
+  };
+  settings: {
+    privacy?: {
+      mode: SpacePrivacyMode;
+    };
   };
 }
