+++ src/main/topLevelPages/myDashboard/ExploreSpaces/ExploreSpacesView.tsx
@@ -12,6 +12,7 @@ import SeeMoreExpandable from '../../../../core/ui/content/SeeMoreExpandable';
 import JourneyTile from '../../../../domain/journey/common/JourneyTile/JourneyTile';
 import { ExploreSpacesViewProps } from './ExploreSpacesTypes';
 import { useColumns } from '../../../../core/ui/grid/GridContext';
+import { SpacePrivacyMode } from '../../../../core/apollo/generated/graphql-schema';
 
 const DEFAULT_ITEMS_LIMIT = 15; // 3 rows of 5 but without the welcome space
 
@@ -61,6 +62,8 @@ export const ExploreSpacesView = ({
   const onFilterChange = (filter: string) => {
     setSelectedFilter(filter);
   };
+
+  const isPrivate = (space): boolean => space?.settings.privacy?.mode === SpacePrivacyMode.Private;
 
   const renderSkeleton = (size: number) =>
     Array.from({ length: size }).map((_, index) => (
@@ -121,7 +124,13 @@ export const ExploreSpacesView = ({
           <>
             {visibleSpaces!.map(space =>
               visibleFirstWelcomeSpace && space.id === welcomeSpace?.id ? null : (
-                <JourneyTile key={space.id} journey={space} journeyTypeName="space" columns={cardColumns} />
+                <JourneyTile
+                  key={space.id}
+                  journey={space}
+                  isPrivate={isPrivate(space)}
+                  journeyTypeName="space"
+                  columns={cardColumns}
+                />
               )
             )}
             {enableLazyLoading && loader}
